refactor: optimized code for yargs
