--- src/utils/isPromise.ts
@@ -1,9 +1,0 @@
-function isPromise(obj: any): obj is PromiseLike<any> {
-  return (
-    !!obj &&
-    (typeof obj === 'object' || typeof obj === 'function') &&
-    typeof obj.then === 'function'
-  );
-}
-
-export default isPromise;
