--- typings/webpack-cli/bin/config/config-yargs.d.ts
@@ -1,6 +1,0 @@
-import yargs from 'yargs';
-
-// eslint-disable-next-line no-shadow
-declare function configYargs<T>(yargs: yargs.Argv<T>): void;
-
-export default configYargs;
